Fix event DTOs type and structure mismatch between order and inventory services

+++ order-service/src/main/java/com/core/orderservice/dto/OrderSummaryDTO.java
@@ -1,6 +1,5 @@
 package com.core.orderservice.dto;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 import com.core.orderservice.domain.OrderStatus;
@@ -10,5 +9,5 @@ public record OrderSummaryDTO(
 		String customerName,
 		Instant orderDate,
 		OrderStatus currentStatus,
-		BigDecimal totalAmount
+		Double totalAmount
 ) {}
